fix double to int

+++ TriangleChecker_20195601A/src/base/TriangleChecker_20195601A.java
@@ -4,47 +4,47 @@ import java.util.*;
 public class TriangleChecker_20195601A {
 	
 	// Validates the input range
-	public static boolean correctRange(double [] input) {
+	public static boolean correctRange(int [] input) {
 		return (input[0] > 0) && (input[1] > 0) && (input[2] > 0);
 	}
 	
 	// Validates the input data type
-	public static boolean collectData(double [] input) {
+	public static boolean collectData(int [] input) {
 		return true;
 	}
 	
 	// Validates the availability  of forming a triangle
-	public static boolean isTri(double [] input) {
+	public static boolean isTri(int [] input) {
 		return (input[0] < input[1] + input[2]) && (input[1] < input[0] + input[2]) && (input[2] < input[0] + input[1]);
 	}
 	
 	// checking if the triangle is an equilateral triangle 
-	public static boolean check_equ_tri(double [] side) {
+	public static boolean check_equ_tri(int [] side) {
 		return ((side[0] == side[1]) && (side[1] == side[2])) && isTri(side) ;
 	}
 	
 	// checking if the triangle is an isosceles triangle
-	public static boolean check_isosceles_tri(double [] side) {
+	public static boolean check_isosceles_tri(int [] side) {
 		return (side[0] == side[1])^(side[0] == side[2])^(side[1] == side[2]) && !(check_equ_tri(side)) && isTri(side);
 	}
 	
 	// checking if the triangle is an scalene triangle
-	public static boolean check_scalene_triangle(double [] side) {
+	public static boolean check_scalene_triangle(int [] side) {
 		return !check_isosceles_tri(side) && !check_equ_tri(side) && isTri(side);
 	}
 	
 	// checking if the triangle is an right angled triangle
-	public static boolean check_right_angled_tri(double [] side) {
-		double a_sqr = side[0]*side[0];
-		double b_sqr = side[1]*side[1];
-		double c_sqr = side[2]*side[2];
+	public static boolean check_right_angled_tri(int [] side) {
+		int a_sqr = side[0]*side[0];
+		int b_sqr = side[1]*side[1];
+		int c_sqr = side[2]*side[2];
 		
 		// boolean isIsosceles = check_isosceles_tri(a_side, b_side, c_side);
 		return ((a_sqr == b_sqr + c_sqr) || (b_sqr == a_sqr + c_sqr) || (c_sqr == a_sqr+b_sqr)) && isTri(side) ;
 	}
 	
-	public static double[] gatherInput(){
-		double [] input = new double[3];
+	public static int[] gatherInput(){
+		int [] input = new int[3];
 		
 		System.out.println("Please enter those 3 sides one by one.");
 		
@@ -52,13 +52,13 @@ public class TriangleChecker_20195601A {
 		
 		try {
 			System.out.println("Enter A side: ");
-			input[0] = sc.nextDouble();
+			input[0] = sc.nextInt();
 			
 			System.out.println("Enter B side: ");
-			input[1] = sc.nextDouble();
+			input[1] = sc.nextInt();
 			
 			System.out.println("Enter C side: ");
-			input[2] = sc.nextDouble();
+			input[2] = sc.nextInt();
 			
 			sc.close();
 		}catch (Exception e) {
@@ -71,7 +71,7 @@ public class TriangleChecker_20195601A {
 	
 	public static void panel() {
 					
-			double[] inputSet = gatherInput();
+			int[] inputSet = gatherInput();
 			
 			if(correctRange(inputSet) && isTri(inputSet)) {
